Pre-load of necessary data before need it
This will make the UI a bit more fluid, hence, a better UX.
